+++ CHANGELOG.md
@@ -2,8 +2,15 @@
 
 ## PENDING
 
+BREAKING CHANGES
+
+FEATURES
+* [lcd] Can now query governance proposals by ProposalStatus
+
 IMPROVEMENTS
 * [baseapp] Allow any alphanumeric character in route
+
+BUG FIXES
 
 ## 0.22.0